+++ src/utils/__tests__/isSchemaViewerEmpty.spec.ts
