+++ src/utils/isExpanded.ts
