+++ src/utils/isSchemaViewerEmpty.ts
